+++ applicatie/src/main/java/com/vdzon/administratie/dto/BestellingRegelDto.java
@@ -5,14 +5,16 @@ import com.vdzon.administratie.model.Bestelling;
 import com.vdzon.administratie.model.BestellingRegel;
 import com.vdzon.administratie.model.FactuurRegel;
 
+import java.math.BigDecimal;
+
 @JsonIgnoreProperties
 public class BestellingRegelDto {
 
     private String uuid;
     private String omschrijving;
-    private double aantal;
-    private double stuksPrijs;
-    private double btwPercentage;
+    private BigDecimal aantal;
+    private BigDecimal stuksPrijs;
+    private BigDecimal btwPercentage;
 
     public BestellingRegelDto() {
     }
@@ -47,15 +49,15 @@ public class BestellingRegelDto {
         return omschrijving;
     }
 
-    public double getAantal() {
+    public BigDecimal getAantal() {
         return aantal;
     }
 
-    public double getStuksPrijs() {
+    public BigDecimal getStuksPrijs() {
         return stuksPrijs;
     }
 
-    public double getBtwPercentage() {
+    public BigDecimal getBtwPercentage() {
         return btwPercentage;
     }
 
@@ -74,9 +76,9 @@ public class BestellingRegelDto {
     public static final class Builder {
         private String uuid;
         private String omschrijving;
-        private double aantal;
-        private double stuksPrijs;
-        private double btwPercentage;
+        private BigDecimal aantal;
+        private BigDecimal stuksPrijs;
+        private BigDecimal btwPercentage;
 
         private Builder() {
         }
@@ -91,17 +93,17 @@ public class BestellingRegelDto {
             return this;
         }
 
-        public Builder aantal(double val) {
+        public Builder aantal(BigDecimal val) {
             aantal = val;
             return this;
         }
 
-        public Builder stuksPrijs(double val) {
+        public Builder stuksPrijs(BigDecimal val) {
             stuksPrijs = val;
             return this;
         }
 
-        public Builder btwPercentage(double val) {
+        public Builder btwPercentage(BigDecimal val) {
             btwPercentage = val;
             return this;
         }
