+++ applicatie/src/main/java/com/vdzon/administratie/model/FactuurRegel.java
@@ -3,15 +3,17 @@ package com.vdzon.administratie.model;
 import org.mongodb.morphia.annotations.Entity;
 import org.mongodb.morphia.annotations.Id;
 
+import java.math.BigDecimal;
+
 @Entity("factuurRegel")
 public class FactuurRegel {
 
     @Id
     private String uuid;
     private String omschrijving;
-    private double aantal;
-    private double stuksPrijs;
-    private double btwPercentage;
+    private BigDecimal aantal;
+    private BigDecimal stuksPrijs;
+    private BigDecimal btwPercentage;
 
     public FactuurRegel() {
     }
@@ -46,24 +48,24 @@ public class FactuurRegel {
         return omschrijving;
     }
 
-    public double getAantal() {
+    public BigDecimal getAantal() {
         return aantal;
     }
 
-    public double getStuksPrijs() {
+    public BigDecimal getStuksPrijs() {
         return stuksPrijs;
     }
 
-    public double getBtwPercentage() {
+    public BigDecimal getBtwPercentage() {
         return btwPercentage;
     }
 
     public static final class Builder {
         private String uuid;
         private String omschrijving;
-        private double aantal;
-        private double stuksPrijs;
-        private double btwPercentage;
+        private BigDecimal aantal;
+        private BigDecimal stuksPrijs;
+        private BigDecimal btwPercentage;
 
         private Builder() {
         }
@@ -78,17 +80,17 @@ public class FactuurRegel {
             return this;
         }
 
-        public Builder aantal(double val) {
+        public Builder aantal(BigDecimal val) {
             aantal = val;
             return this;
         }
 
-        public Builder stuksPrijs(double val) {
+        public Builder stuksPrijs(BigDecimal val) {
             stuksPrijs = val;
             return this;
         }
 
-        public Builder btwPercentage(double val) {
+        public Builder btwPercentage(BigDecimal val) {
             btwPercentage = val;
             return this;
         }
